requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[argon2]
psycopg2-binary
python-dotenv
